Make memory overlap a little less precise so it works with null data ptr
I'm going to make meta tensors have storage (but DataPtr is always
null) so that I can accurately report memory overlap error checking, but
I now have a problem which is that if memory overlap test looks at the
actual data pointer, everything is going to look like it aliases!  A
more conservative test is to just see if the Storage objects themselves
alias, and assume that the data pointers are unique if they don't.

The loss of precision stems from if you unsafely have two distinct
storage objects that point to the same data pointer.  This situation
is pretty rare and so I think it is worth it (and I am hoping no tests
trigger by this.)

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: fd0b2932de3f0998be3bd1bb0d0b61afd556f296
Pull Request resolved: https://github.com/pytorch/pytorch/pull/54710